neo4j improvements on get all method in book repo

+++ BookApp.API/Dtos/BookCatalogListDto.cs
@@ -1,6 +1,6 @@
 namespace BookApp.API.Dtos {
 
   public class BookCatalogListDto {
-    public long CatalogId { get; set; }
+    public int CatalogId { get; set; }
   }
 }
